requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.21.0
requests>=2.25.0
python-dotenv>=0.19.0
pillow>=8.0.0